DB update, book a flight update

+++ VSFly_BerclazFardel/DTO/Booking.cs
@@ -8,6 +8,6 @@ namespace DTO
     {
         public int FlightNo { get; set; }
         public int PassengerID { get; set; }
-        public double SalesPrice { get; set; }
+        public float SalesPrice { get; set; }
     }
 }
