+++ VSFly_BerclazFardel/DTO/Flight.cs
@@ -16,7 +16,7 @@ namespace DTO
 
         public short TotalSeats { get; set; }
 
-        public double BasePrice { get; set; }
+        public float BasePrice { get; set; }
 
         public short SeatsBooked { get; set; }
