+++ VSFly_BerclazFardel/WebAppAPIClient/Models/BookingModel.cs
@@ -8,6 +8,6 @@ namespace WebAppAPIClient.Models
     {
         public int FlightNo { get; set; }
         public int PassengerID { get; set; }
-        public double SalesPrice { get; set; }
+        public float SalesPrice { get; set; }
     }
 }
