+++ VSFly_BerclazFardel/WebAppAPIClient/Models/FlightModel.cs
@@ -16,7 +16,7 @@ namespace WebAppAPIClient.Models
 
         public short TotalSeats { get; set; }
 
-        public double BasePrice { get; set; }
+        public float BasePrice { get; set; }
 
         public short SeatsBooked { get; set; }
     }
